refactor: introduce BaseGuildVoiceChannel class
