+++ src/structures/StageChannel.js
@@ -1,14 +1,12 @@
 'use strict';
 
-const GuildChannel = require('./GuildChannel');
-const Collection = require('../util/Collection');
-const Permissions = require('../util/Permissions');
+const BaseGuildVoiceChannel = require('./BaseGuildVoiceChannel');
 
 /**
  * Represents a guild stage channel on Discord.
  * @extends {GuildChannel}
  */
-class StageChannel extends GuildChannel {
+class StageChannel extends BaseGuildVoiceChannel {
   _patch(data) {
     super._patch(data);
 
@@ -17,59 +15,6 @@ class StageChannel extends GuildChannel {
      * @type {?string}
      */
     this.topic = data.topic;
-
-    /**
-     * The RTC region for this stage channel. This region is automatically selected if `null`.
-     * @type {?string}
-     */
-    this.rtcRegion = data.rtc_region;
-
-    /**
-     * The bitrate of this stage channel
-     * @type {number}
-     */
-    this.bitrate = data.bitrate;
-
-    /**
-     * The maximum amount of users allowed in this channel.
-     * @type {number}
-     */
-    this.userLimit = data.user_limit;
-  }
-
-  /**
-   * The members in this stage channel
-   * @type {Collection<Snowflake, GuildMember>}
-   * @readonly
-   */
-  get members() {
-    const coll = new Collection();
-    for (const state of this.guild.voiceStates.cache.values()) {
-      if (state.channelID === this.id && state.member) {
-        coll.set(state.id, state.member);
-      }
-    }
-    return coll;
-  }
-
-  /**
-   * Checks if the stage channel is full
-   * @type {boolean}
-   * @readonly
-   */
-  get full() {
-    return this.userLimit > 0 && this.members.size >= this.userLimit;
-  }
-
-  /**
-   * Whether the channel is joinable by the client user
-   * @type {boolean}
-   * @readonly
-   */
-  get joinable() {
-    if (!this.viewable) return false;
-    if (!this.permissionsFor(this.client.user).has(Permissions.FLAGS.CONNECT, false)) return false;
-    return true;
   }
 
   /**
@@ -78,37 +23,13 @@ class StageChannel extends GuildChannel {
    * @returns {Promise<StageChannel>}
    * @example
    * // Set the RTC region to europe
-   * stageChannel.setRTCRegion('europe');
+   * voice-basedChannel.setRTCRegion('europe');
    * @example
    * // Remove a fixed region for this channel - let Discord decide automatically
-   * stageChannel.setRTCRegion(null);
+   * voice-basedChannel.setRTCRegion(null);
    */
   setRTCRegion(region) {
     return this.edit({ rtcRegion: region });
-  }
-
-  /**
-   * Attempts to join this stage channel.
-   * @returns {Promise<VoiceConnection>}
-   * @example
-   * // Join a stage channel
-   * stageChannel.join()
-   *   .then(connection => console.log('Connected!'))
-   *   .catch(console.error);
-   */
-  join() {
-    return this.client.voice.joinChannel(this);
-  }
-
-  /**
-   * Leaves this stage channel.
-   * @example
-   * // Leave a stage channel
-   * stageChannel.leave();
-   */
-  leave() {
-    const connection = this.client.voice.connections.get(this.guild.id);
-    if (connection?.channel.id === this.id) connection.disconnect();
   }
 }
